MAP-1402 package and file refactoring
